server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
fonttools==4.46.0
gotrue==2.1.0
h11==0.14.0
holidays==0.38
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lighthouseweb3==0.1.4
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
postgrest==0.13.0
prophet==1.1.5
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
realtime==1.0.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
stanio==0.3.0
starlette==0.27.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.2.0
supafunc==0.3.1
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==11.0.3
